debug hhblits

+++ workflow/graph_construction_hhblits/tmp_annotation.py
@@ -16,30 +16,40 @@ from .hhblits_annotation import (
     Path)
 import logging
 logger=logging.getLogger()
-logging.basicConfig(filename='data/hhblits_annotation.log', 
+logging.basicConfig(filename='data/curated_hhblits_annotation.log', 
                     filemode='a',           
                     level=logging.INFO,
                     format='%(asctime)s - %(levelname)s - %(module)s - %(filename)s - %(message)s')
 def main(infasta:str,cpu:int=4,
-         ori_dir=Path('data/hhblits_ori'),
-         pcs_dir=Path('data/hhblits_pcs')):
+         ori_dir=Path('data/curated_hhblits_ori'),
+         pcs_dir=Path('data/curated_hhblits_pcs')):
     stem=Path(infasta).stem.replace(':genome','')
     
-    try:
-        annotations=hhblits_annotation(infasta,cpu)
-        annotations.to_pickle(ori_dir/f'{stem}.pkl')
-    #     if len(annotations)<1:
-    #         logger.warning(f'BREAK! no hit for "{infasta}"')
-    #         return
-    #     if len(annotations[annotations['probab']>0.8])<1:
-    #         logger.warning(f'no valid hit for "{infasta}"')
-    #         max_probab=annotations['probab'].max()
-    #         logger.warning(f'max_probab: "{max_probab}"')
-    #         return
-    #     grouped_annotations=group_hit(annotations)
-    #     grouped_annotations.to_pickle(pcs_dir/f'{stem}.pkl')
-    except Exception as e:
-        logger.error(f'{stem} failed with error: {e}')
+    if not (ori_dir/f'{stem}.pkl').is_file():
+        annotations=None
+        try:
+            annotations=hhblits_annotation(infasta,cpu)
+            annotations.to_pickle(ori_dir/f'{stem}.pkl')
+            logger.info(f'{stem}: annotated')
+
+        except Exception as e:
+            logger.error(f'{stem}: annotation failed with error:{e}')
+        if annotations is not None:
+            if len(annotations)<1:
+                logger.warning(f'{stem}: no hit')
+                return
+            if len(annotations[annotations['probab']>0.8])<1:
+                max_probab=annotations['probab'].max()
+                logger.warning(f'{stem}: no valid hit, max_probab={max_probab:.2f}')
+                return
+            try:
+                grouped_annotations=group_hit(annotations)
+                grouped_annotations.to_pickle(pcs_dir/f'{stem}.pkl')
+                logger.info(f'{stem}: grouped')
+            except Exception as e:
+                logger.info(f'{stem}: grouped failed with error:{e}')
+    else:
+        logger.info(f'{stem}: processed before, skip')
     # generate_neomodels(infasta,grouped_annotations)
     
 # %%
@@ -49,11 +59,11 @@ if __name__=='__main__':
     from multiprocessing import Pool
     from tqdm import tqdm
     from glob import glob
-    fastas=glob("/home/hugheslab1/zfdeng/pangenome/CoreData/genome_fasta/*:genome.fasta")
+    fastas=glob("/home/hugheslab1/zfdeng/pangenome/CoreData/curated_genome_fasta/*:genome.fasta")[int(sys.argv[1]):int(sys.argv[2])]
     qbar=tqdm(total=len(fastas))
     def callback(i):
         qbar.update()
-    pool=Pool(processes=5,maxtasksperchild=500)
+    pool=Pool(processes=4,maxtasksperchild=500)
     for i in fastas:
         pool.apply_async(main,(i,),callback=callback)
     pool.close()
